legacy/docs/whiteblock/test_plan_automation/requirements.txt
pyyaml>=4.2b1
optparse-pretty==0.1.1